llm-srv/pyproject.toml
[project]
name = "llm-srv"
version = "0.1.0"
description = ""
authors = [
    {name = "Kye Burchard",email = "kyeb@mit.edu"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.1"
]

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
